rename parser package, rename python files in vector_db
