+++ parsers/requirements.txt
@@ -1,4 +1,2 @@
 pymupdf==1.24.4
 langchain==0.2.1
-loguru==0.7.2
-unstructured[pdf]
