+++ vector_db/start.py
@@ -1,14 +1,18 @@
 import os
-import logging
+import time
+import subprocess
 
 import chromadb
 from chromadb.api import ClientAPI
 from chromadb.api.models import Collection
 from dotenv import load_dotenv
+from loguru import logger
 
-# Set up logging
-logging.basicConfig(level=logging.INFO)
-logger = logging.getLogger(__name__)
+
+def start_chroma_db_server(host: str, port: int, db_path: str) -> subprocess.Popen:
+    cmd = f"chroma run --host {host} --port {port} --path {db_path}"
+    process = subprocess.Popen(cmd, shell=True)
+    return process
 
 
 def get_client(host: str, port: int) -> ClientAPI:
@@ -28,11 +32,23 @@ def get_collection(collection_name: str, host: str = 'localhost', port: int = 80
 
 
 if __name__ == "__main__":
-    # command line: chroma run --host localhost --port 8000 --path ../vectordb-stores/chromadb
     # Load environment variables from .env file
     load_dotenv()
     _host = os.getenv("CHROMA_HOST", "localhost")
     _port = int(os.getenv("CHROMA_PORT", 8000))
     _collection_name = os.getenv("COLLECTION_NAME", "default_collection")
-    # Initialize the ChromaDB collection
-    _ = get_collection(_collection_name, _host, _port)
+    _db_path = os.getenv("CHROMA_DB_PATH", "../vectordb-stores/chromadb")
+    # Start the ChromaDB server
+    chroma_process = start_chroma_db_server(_host, _port, _db_path)
+    # Wait for a while for the server to start
+    time.sleep(10)
+    try:
+        # Initialize the ChromaDB collection
+        _ = get_collection(_collection_name, _host, _port)
+        logger.info("Chroma DB started")
+    finally:
+        # Ensure the ChromaDB server is stopped after initialization
+        chroma_process.terminate()
+        chroma_process.wait()
+        logger.info("Chroma DB terminated")
+
